add support for dnf package manager

+++ install.sh
@@ -1,12 +1,19 @@
 # https://github.com/flowintel/flowintel
-sudo apt install -y python3-venv git screen libolm-dev librsvg2-bin wget
 
-# install pandoc from git 
-# pandoc dependencies
-sudo apt install -y texlive texlive-xetex texlive-fonts-extra
-wget https://github.com/jgm/pandoc/releases/download/3.7/pandoc-3.7-1-$(dpkg --print-architecture).deb
-sudo dpkg -i pandoc*.deb
-rm pandoc*
+if apt -v &> /dev/null ; then
+    sudo apt install -y python3-venv git screen libolm-dev librsvg2-bin wget
+    # install pandoc from git 
+    # pandoc dependencies
+    sudo apt install -y texlive texlive-xetex texlive-fonts-extra
+    wget https://github.com/jgm/pandoc/releases/download/3.7/pandoc-3.7-1-$(dpkg --print-architecture).deb
+    sudo dpkg -i pandoc*.deb
+    rm pandoc*
+elif dnf --version &> /dev/null ; then
+    #RockyLinux
+    sudo dnf install -y epel-release
+    sudo crb enable
+    sudo dnf in -y pandoc python3 git screen libolm librsvg2 wget
+fi
 
 # Install a template for the export of notes in pdf
